backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
branca==0.7.2
build==1.2.1
cachetools==5.3.3
cartes==0.7.7
Cartopy==0.23.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
crewai==0.30.11
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.6
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docstring-parser==0.15
email_validator==2.1.1
embedchain==0.1.103
evdev==1.7.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fiona==1.9.6
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
geopandas==0.14.4
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
groq==0.8.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
impunity==1.0.4
instructor==0.5.2
ipyleaflet==0.19.1
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-leaflet==0.19.1
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.56
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
metar==1.11.0
minio==7.2.7
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.9
onnxruntime==1.17.3
openai==1.30.1
openap==1.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
Pint==0.23
pitot==0.3.1
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pynput==1.7.7
pyopensky==2.6
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.1.0
pysbd==0.3.4
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rs1090==0.2.0
rsa==4.9
schema==0.7.7
scipy==1.13.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tiktoken==0.5.2
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.4
traffic==2.10.0
traitlets==5.14.3
traittypes==0.2.1
trino==0.328.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.1
